Move some of misc1.c to indent.c.
